pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "authora"
version = "0.2.0"
description = "Official Python SDK for the Authora agent identity and authorization platform."
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Authora", email = "sdk@authora.dev" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "httpx>=0.25.0,<1.0.0",
    "pynacl>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "respx>=0.21",
    "mypy>=1.5",
    "ruff>=0.1",
]

[project.urls]
Homepage = "https://authora.dev"
Documentation = "https://authora.dev/developers/sdk/python"
Repository = "https://github.com/authora-dev/authora-python"
Issues = "https://github.com/authora-dev/authora-python/issues"
Changelog = "https://authora.dev/developers/changelog"

[tool.hatch.build.targets.wheel]
packages = ["authora"]

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
target-version = "py39"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
